refactored tidy_graph to add tidy&untidy output
the tidy_graph function now lives on the ControlFlowGraph rather
than being privately called inside the cfg::Builder. this is just
so that during debugging an untidied AND tidied variant can be
emitted as a .dot file for viewing with graphviz.

also added an exit node to functions control flow graph, automatically
generated if a function ends in a node with no exit edges.

minor changes were made to the way Block's Debug implementation
is done so that the graphviz views look prettier, very important.